backend/requirements.txt
Django==4.2.25
djangorestframework==3.15.2
django-filter==25.1
djoser==2.1.0
pillow==10.0.1
python-dotenv==1.2.1
psycopg2-binary==2.9.11
gunicorn==23.0.0
flake8==7.3.0
flake8-isort==6.1.2
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest-lazy-fixture==0.6.3
hashids
PyYAML
